Fix static analysis findings in RedisTest
- Move the fake Credis_Client stand-in into its own file
  (RedisTestFakeClient.php): the coding standard requires one class
  per file.
- Add missing @var docblocks on two of its properties.
- Drop the unused $field parameter from its hGet() - the fake never
  needed it, since results are keyed by id alone.
