+++ lib/internal/Magento/Framework/Cache/Test/Unit/Backend/RedisTest.php
@@ -119,65 +119,3 @@ class RedisTest extends TestCase
         );
     }
 }
-
-/**
- * Minimal stand-in for Credis_Client covering only what Redis::load()/save()/remove() call
- * directly: a pipelined batch of hGet() calls flushed by exec(), and a standalone hGet() for the
- * non-preloaded path. Anything save()/remove() need beyond that (hMSet, multi, del, ...) is left
- * unimplemented on purpose, so those calls surface as an Error that the caller's own try/catch
- * around parent::save()/parent::remove() is expected to swallow.
- */
-class RedisTestFakeClient
-{
-    public int $pipelineCount = 0;
-
-    /** @var array<int, array<int, mixed>> */
-    private array $queuedExecResults = [];
-
-    /** @var array<string, mixed> */
-    private array $directResults = [];
-
-    private bool $inPipeline = false;
-
-    public function queueExecResult(array $result): void
-    {
-        $this->queuedExecResults[] = $result;
-    }
-
-    public function setDirectResult(string $id, $value): void
-    {
-        $this->directResults[$id] = $value;
-    }
-
-    public function pipeline(): self
-    {
-        $this->pipelineCount++;
-        $this->inPipeline = true;
-
-        return $this;
-    }
-
-    public function hGet(string $key, string $field)
-    {
-        if ($this->inPipeline) {
-            // Queued: the actual value is returned positionally by exec(), not here.
-            return $this;
-        }
-
-        $id = $this->stripKeyPrefix($key);
-
-        return $this->directResults[$id] ?? false;
-    }
-
-    public function exec(): array
-    {
-        $this->inPipeline = false;
-
-        return array_shift($this->queuedExecResults) ?? [];
-    }
-
-    private function stripKeyPrefix(string $key): string
-    {
-        return substr($key, strlen(Redis::PREFIX_KEY));
-    }
-}
